Implement basic contract + web3 functionality
New file taxicoin.js is used for interacting with web3 and the contract.
Currently contains mostly method stubs, with a couple of examples of
interacting with a contract instance.

Includes a temporary fix for the build pipeline where contracts could
not be deployed from a custom build dir. Now builds to and deploys from
default dir, and then copies artifacts to custom dir.